refactor: Refactor MessageRepositoryClient without nestjs structure
