+++ packages/client/main.ts
@@ -1,0 +1,1 @@
+export { MessagePickupRepositoryClient } from './MessagePickupRepositoryClient'
