--- packages/client/src/client.module.ts
@@ -1,9 +1,0 @@
-import { Module } from '@nestjs/common'
-import { MessagePickupRepositoryClient } from './lib/MessagePickupRepositoryClient'
-
-@Module({
-  imports: [],
-  controllers: [],
-  providers: [MessagePickupRepositoryClient],
-})
-export class ClientModule {}
