--- packages/client/test/app.e2e-spec.ts
@@ -1,21 +1,0 @@
-import { Test, TestingModule } from '@nestjs/testing'
-import { INestApplication } from '@nestjs/common'
-import * as request from 'supertest'
-import { ClientModule } from './../src/client.module'
-
-describe('ClientController (e2e)', () => {
-  let app: INestApplication
-
-  beforeEach(async () => {
-    const moduleFixture: TestingModule = await Test.createTestingModule({
-      imports: [ClientModule],
-    }).compile()
-
-    app = moduleFixture.createNestApplication()
-    await app.init()
-  })
-
-  it('/ (GET)', () => {
-    return request(app.getHttpServer()).get('/').expect(200).expect('Hello World!')
-  })
-})
