+++ tsconfig.json
@@ -18,7 +18,7 @@
     "forceConsistentCasingInFileNames": false,
     "noFallthroughCasesInSwitch": false,
     "paths": {
-      "@message-pickup-repository/client": ["packages/client/src/lib/MessagePickupRepositoryClient.ts"]
+      "@message-pickup-repository/client": ["packages/client/MessagePickupRepositoryClient.ts"]
     }
   },
   "include": ["src/**/*.ts", "packages/**/*.ts"]
